+++ students/heesu/users/urls.py
@@ -2,5 +2,6 @@ from django.urls import path
 from users.views import UserView
 
 urlpatterns = [
-    path('', UserView.as_view()),
+    path('/signup', UserView.as_view()),
+    path('/sign', UserView.as_view()),
 ]
